Okay, but now fix it so that styles actually work
A bit more involved, since there's not a lot we can do to get around the style manager silliness. This is a space compromise, as well: now, styles will get IDs based on their object references, so if you use multiple duplicate objects for the same style, stuff will get inflated (might introduce a factory to work again this later). That said, it basically still accomplishes what we needed.