Image support; mostly implemented cell tooltip support
...I started the tooltip stuff ages ago and never actually finished it. I think it's mostly working, though, and will hopefully get around to hooking it up to something.

As for the image support, it should be stable -- the only limitation is you have to specify the cell style as image, and then have the cell value be URL (keeps things _a lot_ cleaner). It should be pretty fast -- we use HAMR's proxy to cache and speed things up, and we will query up to _20_ images simultaneously. In other words, this is about as fast as this kind-of thing can get.

(The added image functionality should not have a substantial performance impact on other exports. We do have to scan every row for images to set the row height now, but I don't believe this will cause a substantial slowdown.)

...Oh, I should probably clean this up still. Maybe later.