Optimise some stuff by breaking upstream compatability
This might break quite a few things, in truth, but I don't believe it breaks much as far as HAMR is concerned. The changes here are not _major_, but they seem capable of a 3-5% speed increase for SKU portfolio generation, and I'm going to take any improvements I can find.